03_production_system/pyproject.toml
[project]
name = "income_prediction_agent"
version = "0.1.0"
description = "income_prediction_agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai>=0.165.1,<1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0", 
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "xgboost>=2.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
income_prediction_agent = "income_prediction_agent.main:run"
run_crew = "income_prediction_agent.main:run"
train = "income_prediction_agent.main:train"
replay = "income_prediction_agent.main:replay"
test = "income_prediction_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/income_prediction_agent"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.crewai]
type = "crew"